[REFACTOR] move initialState to separate a file
