+++ src/initialState.js
@@ -1,0 +1,9 @@
+const initialState = {
+  regions: [],
+  categories: [],
+  selectedRegion: [],
+  selectedCategory: [],
+  restaurants: [],
+};
+
+export default initialState;
